requirements.txt
# DEPENDÊNCIAS PARA RODAR O PROJETO:

fastapi
uvicorn
jinja2
python-multipart
pdfplumber
ollama
openai
google-genai
python-dotenv
